pyproject.toml
# Defines the build system used for the project.
[build-system]
requires = ["hatchling>=1,<2"]
build-backend = "hatchling.build"

# Project metadata section. Provides the general ID information about the project.
[project]
name = "sollertia-shared-assets"
version = "8.0.0rc5"
description = "Provides data acquisition and processing assets shared between Sollertia platform libraries."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.14,<3.15"
authors = [
    { name = "Ivan Kondratyev" },
    { name = "Kushaan Gupta" },
    { name = "Natalie Yeung" },
]
maintainers = [
    { name = "Ivan Kondratyev", email = "ik278@cornell.edu" },
]

keywords = ["sollertia", "ataraxis", "dataclasses", "configuration"]

classifiers = [
    # Development status
    "Development Status :: 5 - Production/Stable",
    # Intended audience and project topic
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering",
    # Supported Python Versions
    "Programming Language :: Python :: 3.14",
    # Supported Operating Systems
    "Operating System :: OS Independent",
    # Typing
    "Typing :: Typed",
]

# Runtime project dependencies.
dependencies = [
    # Ataraxis Libraries
    "ataraxis-time>=6,<7",
    "ataraxis-base-utilities>=6,<7",
    "ataraxis-data-structures>=6,<7",

    # Configuration and CLI
    "platformdirs>=4,<5",
    "click>=8,<9",

    # Date parsing (used by session-discovery filter helpers)
    "python-dateutil>=2,<3",

    # Data Processing
    "polars>=1,<2",
    "pyarrow>=23,<24",

    # MCP and HTTP
    "mcp>=1,<2",
    "httpx>=0.28,<1",
]

[project.urls]
Homepage = "https://github.com/Sun-Lab-NBB/sollertia-shared-assets"
Documentation = "https://sollertia-shared-assets-api-docs.netlify.app/"

# Exports the CLI for interacting with all interactive components of the sollertia-shared-assets library.
[project.scripts]
slsa = "sollertia_shared_assets.interfaces.cli:slsa_cli"

# Development dependencies (PEP 735). Not published to PyPI.
[dependency-groups]
dev = [
    # Tox
    "tox>=4,<5",
    "uv>=0,<1",
    "tox-uv>=1,<2",

    # Development Automation
    "ataraxis-automation>=8,<9",

    # Type stubs
    "types-python-dateutil>=2,<3",
]

# Specifies files that should not be included in the source-code distribution but are also not
# part of gitignore.
[tool.hatch.build.targets.sdist]
exclude = [".github", "recipe"]

# Specifies the library structure.
[tool.hatch.build.targets.wheel]
packages = ["src/sollertia_shared_assets"]

# Ruff Configuration.
[tool.ruff]
line-length = 120         # The maximum column length is set to 120 for this project.
indent-width = 4          # Same as black, indents are 4 spaces
target-version = "py314"  # Targets the lowest supported version of python
src = ["src"]             # The name of the source directory

# Excludes 'service' .py files, such as the sphinx configuration file, from linting.
extend-exclude = ["conf.py"]

# Checks for all potential violations and uses the exclusions below to target-disable specific ones.
lint.select = ["ALL"]

# General ignore directives
lint.ignore = [
    "W291",    # Conflicts with docstring formatting
    "TID252",  # Relative imports are important for correct asset management in this library
    "BLE001",  # Blind except needed for robust error handling in data transfer and MCP tools
    "C901",    # Sometimes complex functions are necessary
    "PLR0913", # Sometimes complex functions are necessary
    "PLR0912", # Sometimes complex functions are necessary
    "PLR0915", # Sometimes complex type definitions are necessary
    "PLR0911", # Sometimes complex return graphs are necessary
    "COM812",  # Conflicts with the formatter
    "ISC001",  # Conflicts with the formatter
    "PT001",   # https://github.com/astral-sh/ruff/issues/8796#issuecomment-182590771
    "PT023",   # https://github.com/astral-sh/ruff/issues/8796#issuecomment-1825907715
    "D107",    # __init__ is documented inside the main class docstring where applicable
    "D205",    # Bugs out for file descriptions
    "PLW0603", # While global statement usage is not ideal, it streamlines certain development patterns
]

# Additional formatting configurations
[tool.ruff.format]
quote-style = "double"             # Uses double quotes for strings
indent-style = "space"             # Uses space for indents
skip-magic-trailing-comma = false  # Like black, ignores trailing commas
line-ending = "auto"               # Automatically detects and standardizes line-ending character

# Docstrings and comments' line length
[tool.ruff.lint.pycodestyle]
max-doc-length = 120  # Maximum documentation line length, the same as code

# Docstrings style
[tool.ruff.lint.pydocstyle]
convention = "google"

# Additional, file-specific 'ignore' directives
[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
    "F401", # Imported but unused
    "F403", # Wildcard imports
]

[tool.ruff.lint.isort]
case-sensitive = true              # Takes case into account when sorting imports
combine-as-imports = true          # Combines multiple "as" imports for the same package
force-wrap-aliases = true          # Wraps "as" imports so that each uses a separate line (after combining for package)
force-sort-within-sections = true  # Forces "as" and "from" imports for the same package to be close
length-sort = true                 # Places shorter imports first

# MyPy configuration section.
[tool.mypy]
disallow_untyped_defs = true # Enforces function annotation
warn_unused_ignores = true   # Warns against using 'type: ignore' for packages that have type stubs available.
exclude = [
    "sollertia-shared-assets-\\d+", # Ignores temporary folder created by setuptools when building the sdist
    "venv.*/",               # Ignores virtual environments
    "build/",                # Ignores the sdist directory
    "dist/",                 # Ignores the wheel directory
    "docs/",                 # Ignores the sphinx / doxygen directory
    "stubs/",                # Ignores stubs directory (stubgen output target)
    "recipe/",               # Ignores recipe directory (grayskull output target)
    "tests/",                # Ignores the test directory.
]

# This is used by the 'test' tox tasks to aggregate coverage data produced during pytest runtimes.
[tool.coverage.run]
omit = ["*/interfaces/*"]

[tool.coverage.paths]

# Maps coverage measured in site-packages to source files in src
source = ["src/", ".tox/*/lib/python*/site-packages/"]

# Same as above, specifies the output directory for the coverage .html report
[tool.coverage.html]
directory = "reports/coverage_html"

# Specifies additional ignore directives
[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __del__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]
